+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+## Fixed
+- Fixed url methods return type when a part is not found
+
 ## [1.1.4] - 2017-06-28
 ## Added
 - Pass component to `setResource()` callback